Rollback try 3

+++ submission/policy/frontier_exploration_policy.py
@@ -117,18 +117,18 @@ class FrontierExplorationPolicy(Policy):
                 if goal_category_cpu[e] == 1:
                     # If we're looking for a couch, filter out all cells that
                     # also have been classified as a chair or a bed
-                    category_map -= (map_features[e, 0 + 8, :, :] == 1)
-                    category_map -= (map_features[e, 3 + 8, :, :] == 1)
+                    category_map -= (map_features[e, 0 + 8, :, :] == 1).float()
+                    category_map -= (map_features[e, 3 + 8, :, :] == 1).float()
 
                 elif goal_category_cpu[e] == 3:
                     # If we're looking for a bed, filter out couch
-                    category_map -= (map_features[e, 1 + 8, :, :] == 1)
+                    category_map -= (map_features[e, 1 + 8, :, :] == 1).float()
 
                 elif goal_category_cpu[e] == 0:
                     # If we're looking for a chair, filter out couch and
                     # bed (frame)
-                    category_map -= (map_features[e, 1 + 8, :, :] == 1)
-                    category_map -= (map_features[e, 3 + 8, :, :] == 1)
+                    category_map -= (map_features[e, 1 + 8, :, :] == 1).float()
+                    category_map -= (map_features[e, 3 + 8, :, :] == 1).float()
 
                 if goal_category_cpu[e] in [0, 1, 3]:
                     # For large objects (chair, couch, bed), remove noise with
